Correct behavior for widening range on open boundaries
Critically, expanding by only 4σ in each direction is insufficient
since the convolution kernel is symmetric and has a standard deviation
of σ; therefore, the fully width of the truncated Gaussian is 8σ.

This wider range also made the suboptimal ordering in widening the
range and calculating the number of bins (when not explicitly given)
more obvious since a few bins were being shared by a larger span of
the axis. The goal is to choose a number of bins appropriate to the
resolution of the data, so the `nbins` estimate should happen after
any necessary expansion of the axis range.