Add an implementation of the Improved Sheather-Jones bandwidth estimator
This adds the ISJ bandwidth estimator described in Botev et al and
uses it by default, as it is more capable of dealing both with
non-Gaussian distributions and respects the complexity/nature of
bounded domains.

Z. Botev, J. Grotowski and D. Kroese. "Kernel density estimation via
diffusion." The Annals of Statistics 38 (2010), arXiv:1011.2602.