Add new interface for calculating KDE interval
Add a new `bounds` argument, and define the set of
`lo`, `hi`, and `boundary` arguments to be a convenience for setting
`bounds = (lo, hi, boundary)`.

The new `bounds()` interface method, corresponding to the eponymous
keyword argument, allows for extending the package to support more
sophisticated bounds determination. For instance, this is a step
towards supporting automatic interval and boundary condition
determination for distributions from `Distributions` (which would make
handling KDEs of MCMC chains given a specific prior easy).