Implement more accurate histogram binning calculation
With the naive calculation used prior to this commit, the limits of
finite floating point means we can miss a bin edge by epsilon, leading
to getting the bin weights very wrong. Enhance the bin index calculation
by explicitly calculating the truncated value in the rescaling factor,
and adjust each fractional bin index with that residual (using a couple
of "double-double" steps).

(This is a much more performant version compared to switching to using
`searchsortedlast` on the bin `edges` vector, but _should_ give
essentially the same answers. In contrast to the benchmarks below,
the range-based search regressed by roughly a factor of 4×.)

There is **no** significant change in performance:

```julia
import KernelDensityEstimation as KDE
using BenchmarkTools

rv = randn(1_000_000);
@benchmark KDE.kde($rv, method = KDE.LinearBinning(), bwratio = 1.0)
```

**Before**
```
BenchmarkTools.Trial: 750 samples with 1 evaluation.
 Range (min … max):  6.452 ms …   7.639 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.582 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.670 ms ± 244.854 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▂▄▆▇██▇▆▅▅▂▂▂▁ ▁
  █████████████████▅█▅▇▄▅▇▇▄▅▆▆▆▄▅▁▁▄▆▁▅▅▁▁▅▁▇▆▅▄▅▅▅▆▅▇▅▇▅█▇▆ █
  6.45 ms      Histogram: log(frequency) by time      7.54 ms <

 Memory estimate: 8.28 KiB, allocs estimate: 54.
```

**After**
```
BenchmarkTools.Trial: 815 samples with 1 evaluation.
 Range (min … max):  5.940 ms …  6.849 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.128 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.135 ms ± 95.099 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▁▄ ▃▄▆▃▅▂▂█▄ ▄ ▅▄ ▄▃▄▄▃▃▂▆ ▁ ▁▁
  ▂▄▁▃▄▆▅▅▇████████████▇██████████████████▅▇▅▆▄▅▃▂▃▃▃▂▃▃▂▂▄▃ ▅
  5.94 ms        Histogram: frequency by time        6.39 ms <

 Memory estimate: 8.28 KiB, allocs estimate: 54.
```