Add support for weighted samples
Add a new optional keyword argument in `kde` to support passing
arbitrary weights to the underlying implementations. This requires
changing the interfaces for the internal `estimate` and `bandwidth`
functions to support the additional data, and the binning and bandwidth
estimation algorithms have all been updated to make use of the weights.