Fix infrequent bad KDE due to negative values in density
The function to force positivity in the linear boundary correction
case only works when the basic KDE is strictly non-negative. This is
always true mathematically (for a non-negative kernel), but numerically
the convolution does result in very small negative values.

For very particular inputs, this violated assumption leads to a very
large positive exponent in the `pos()` function at a single point, so
the KDE ends up dominated by a single spike.

Enforce the desired invariant by assuming all negative values are
really just zeros.

I've been unable to construct a small, isolated test case for this
behavior since it is so sensistive to inputs, optimizations, etc. that
all conspire to a paricular pattern of small positive/negative numbers
hovering around zero.