Change data representation required by histogramming algorithm
While it was attractive on the surface to permit a simple `reinterpret`
of a data matrix as a vector of coordinate tuples, this was a very
restrictive requirement for non-bitstype data types. Instead, switch
to taking in a tuple of vectors.

(The matrix case can still be made essentially copy-free by using
`view`s of rows or columns in a tuple.)