Consolidate limits and boundary condition interface into single function
The `bounds` function interface now handles both inferring the (finite)
span of the KDE as well as the boundary condition of the interval,
rather than separating these two steps into separate functions. This
reduces the number of interfaces that need to be implemented and should
make the dance to handle multivariate inputs easier.